pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-travel-agent"
version = "1.0.0"
description = "AI-powered travel planning application with Flask"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AI Travel Agent Team", email = "contact@aitravelagent.com"}
]
maintainers = [
    {name = "AI Travel Agent Team", email = "contact@aitravelagent.com"}
]
keywords = ["travel", "ai", "flask", "openai", "planning", "tourism"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
requires-python = ">=3.8"
dependencies = [
    "flask>=2.3.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.29",
    "langchain-experimental>=0.0.56",
    "langgraph>=0.0.45",
    "python-dotenv>=1.0.0",
    "google-search-results>=2.4.2",
    "duckduckgo-search>=4.0.0",
    "requests>=2.31.0",
    "pyowm>=3.3.0",
    "youtube-search-python>=1.6.6",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]

[project.scripts]
ai-travel-agent = "app:main"

[project.urls]
Homepage = "https://github.com/yourusername/ai-travel-agent"
Documentation = "https://github.com/yourusername/ai-travel-agent#readme"
Repository = "https://github.com/yourusername/ai-travel-agent.git"
Issues = "https://github.com/yourusername/ai-travel-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
"*" = ["templates/*.html", "static/*.css", "static/*.js"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
